New dockerfile for coref
