+++ Dockerfile
@@ -26,7 +26,7 @@ COPY tests/ tests/
 COPY app.py app.py
 COPY server/ server/
 
-RUN pytest tests/
+RUN pytest tests/general_tests/
 
 # Copy the configuration files used at runtime
 COPY models.json models.json
