+++ coref_requirements.txt
@@ -1,0 +1,14 @@
+psycopg2-binary
+python-json-logger
+pytorch-pretrained-bert
+
+# REST interface for models
+flask>=1.0.2
+flask-cors>=3.0.7
+gevent>=1.3.6
+
+# For timezone utilities
+pytz>=2017.3
+
+# Pinned because werkzeug.contrib.fixers is deprecated in 1.0.
+werkzeug<=0.16.1
